+++ components/level/tile/player.tsx
@@ -2,8 +2,8 @@ import { AppContext } from '@root/contexts/appContext';
 import { GridContext } from '@root/contexts/gridContext';
 import classNames from 'classnames';
 import React, { useContext } from 'react';
-import Theme, { getIconFromTheme } from '../../constants/theme';
-import TileType from '../../constants/tileType';
+import Theme, { getIconFromTheme } from '../../../constants/theme';
+import TileType from '../../../constants/tileType';
 import styles from './Player.module.css';
 
 interface PlayerProps {
